+++ version/changelog.md
@@ -2,6 +2,10 @@
 
 ## changelog
 
+###
+
+- navigate to parent folder automatically if folder does not exist (any more)
+
 ### v1.0.0 (25.01.2026)
 
 - base app with file commander utility